+++ Realm/RLMSyncConfiguration_Private.hpp
@@ -18,14 +18,31 @@
 
 #import "RLMSyncConfiguration_Private.h"
 
+#import <functional>
+#import <memory>
+
 namespace realm {
+class SyncSession;
 struct SyncConfig;
+struct SyncError;
+using SyncSessionErrorHandler = void(std::shared_ptr<SyncSession>, SyncError);
 }
 
+NS_ASSUME_NONNULL_BEGIN
+
 @interface RLMSyncConfiguration ()
+
+- (instancetype)initWithUser:(RLMSyncUser *)user
+                    realmURL:(NSURL *)url
+               customFileURL:(nullable NSURL *)customFileURL
+                   isPartial:(BOOL)isPartial
+                  stopPolicy:(RLMSyncStopPolicy)stopPolicy
+                errorHandler:(std::function<realm::SyncSessionErrorHandler>)errorHandler;
 
 - (instancetype)initWithRawConfig:(realm::SyncConfig)config;
 
 - (realm::SyncConfig)rawConfiguration;
 
 @end
+
+NS_ASSUME_NONNULL_END