+++ Realm/RLMSyncUser.mm
@@ -24,6 +24,8 @@
 #import "RLMRealmConfiguration_Private.hpp"
 #import "RLMRealmUtil.hpp"
 #import "RLMResults_Private.hpp"
+#import "RLMSyncConfiguration.h"
+#import "RLMSyncConfiguration_Private.hpp"
 #import "RLMSyncManager_Private.h"
 #import "RLMSyncPermissionResults.h"
 #import "RLMSyncPermission_Private.hpp"
@@ -156,11 +158,9 @@ PermissionChangeCallback RLMWrapPermissionStatusCallback(RLMPermissionStatusBloc
 - (instancetype)initPrivate {
     if (self = [super init]) {
         _configMaker = std::make_unique<ConfigMaker>([](std::shared_ptr<SyncUser> user, std::string url) {
-            RLMRealmConfiguration *config = [RLMRealmConfiguration defaultConfiguration];
             NSURL *objCUrl = [NSURL URLWithString:@(url.c_str())];
             RLMSyncUser *objCUser = [[RLMSyncUser alloc] initWithSyncUser:std::move(user)];
-            config.syncConfiguration = [[RLMSyncConfiguration alloc] initWithUser:objCUser realmURL:objCUrl];
-            return [config config];
+            return [objCUser configurationWithURL:objCUrl fullSynchronization:true].config;
         });
         return self;
     }
@@ -204,6 +204,44 @@ PermissionChangeCallback RLMWrapPermissionStatusCallback(RLMPermissionStatusBloc
                               timeout:timeout
                         callbackQueue:callbackQueue
                       completionBlock:completion];
+}
+
+- (RLMRealmConfiguration *)configuration {
+    return [self configurationWithURL:nil
+                  fullSynchronization:NO
+                  enableSSLValidation:YES
+                            urlPrefix:nil];
+}
+
+- (RLMRealmConfiguration *)configurationWithURL:(NSURL *)url {
+    return [self configurationWithURL:url
+                  fullSynchronization:NO
+                  enableSSLValidation:YES
+                            urlPrefix:nil];
+}
+
+- (RLMRealmConfiguration *)configurationWithURL:(NSURL *)url fullSynchronization:(bool)fullSynchronization {
+    return [self configurationWithURL:url
+                  fullSynchronization:fullSynchronization
+                  enableSSLValidation:YES
+                            urlPrefix:nil];
+}
+
+- (RLMRealmConfiguration *)configurationWithURL:(NSURL *)url
+                            fullSynchronization:(bool)fullSynchronization
+                            enableSSLValidation:(bool)enableSSLValidation
+                                      urlPrefix:(NSString * _Nullable)urlPrefix {
+    auto syncConfig = [[RLMSyncConfiguration alloc] initWithUser:self
+                                                        realmURL:url ?: self.defaultRealmURL
+                                                   customFileURL:nil
+                                                       isPartial:!fullSynchronization
+                                                      stopPolicy:RLMSyncStopPolicyAfterChangesUploaded
+                                                    errorHandler:nullptr];
+    syncConfig.urlPrefix = urlPrefix;
+    syncConfig.enableSSLValidation = enableSSLValidation;
+    RLMRealmConfiguration *config = [[RLMRealmConfiguration alloc] init];
+    config.syncConfiguration = syncConfig;
+    return config;
 }
 
 - (void)logOut {